Delete Msg Confirm Added

+++ adminPanel/adminBillSearch.php
@@ -64,7 +64,7 @@
                             <td><?php echo $i['total'];?></td>
                             <td><?php echo $i['_time'];?></td>
                             <td><a class="btn btn-primary" href="../adminPanel/adminBillUpdate.php?id=<?php echo $fid.'/'. $i['_time'];?>" role="button">Update Bill</a></td>
-                            <td><a class="btn btn-danger" href="../adminPanel/adminBillDelete.php?id=<?php echo $fid.'/'. $i['_time'];?>" role="button">Delete Bill</a></td>
+                            <td><a class="btn btn-danger" onclick="javascript:return confirm('Are you sure to delete this??');" href="../adminPanel/adminBillDelete.php?id=<?php echo $fid.'/'. $i['_time'];?>" role="button">Delete Bill</a></td>
                         </tr>
                 <?php } ?>
             </tbody>
